Move entry construction method
