+++ Packages/EventsWidgetLogic/Sources/EventsWidgetLogic/Adapting Events/EventCluster.swift
@@ -1,0 +1,9 @@
+import Foundation.NSDate
+
+struct EventCluster {
+    
+    var clusterStartTime: Date
+    var events: [Event]
+    var additionalEventCount: Int
+    
+}
