+++ Packages/EventsWidgetLogic/Sources/EventsWidgetLogic/Adapting Events/EventTimelineEntry.swift
@@ -12,4 +12,13 @@ public struct EventTimelineEntry: Equatable {
         self.additionalEventsCount = additionalEventsCount
     }
     
+    init(cluster: EventCluster) {
+        let viewModels = cluster.events.map(EventViewModel.init)
+        self.init(
+            date: cluster.clusterStartTime,
+            events: viewModels,
+            additionalEventsCount: cluster.additionalEventCount
+        )
+    }
+    
 }
